License and documentation were not included in the distribution packages. They are added to the outer Makefile.am as EXTRA_DIST.
-lm flags were missing in AM_LDFLAGS, and caused compilation error in
Ubuntu and GhostBSD. The math library is now linked in both interpreter source and test
sources.

The program did not compile in very old Mac OS X Leopard on
PPC. In newer versions of GCC, redefinitions of typedef with
compatible types is allowed, but old versions produce error.
Redefinitions were unintentional and unnecessary. They are removed.

In old GCC, language standard is C89 by default which leads to error
because of the use of variable declarations in for loops.
AM_PROG_CC_C99 is added to configure.ac to solve this.

The uses of inline were wrong and they produced errors when
optimization level gets changed. All uses of inline keyword are removed.

In scanner, buffer sizes were limited to 128 characters, and
there was no mechanism to check whether an overflow happens. I forgot
to fix this problem for a long time. Now all buffers sizes are
dynamic and unlimited.

constexpr and define-constexpr are added to syntax.tl.

Documentation is updated to include macros. It is still incomplete.